+++ source/scan-tool-cache/main.cpp
@@ -23,7 +23,9 @@
 #include <string>
 #include "../../libthoro/common/StringUtils.hpp"
 #include "../../libthoro/filesystem/directory.hpp"
+#include "../../libthoro/filesystem/file.hpp"
 #include "../virustotal/CacheManagerV2.hpp"
+#include "../Configuration.hpp"
 #include "../Constants.hpp"
 //return codes
 #include "../ReturnCodes.hpp"
@@ -56,6 +58,10 @@ void showHelp()
             << "                     report or initiating a rescan. This operation requires an\n"
             << "                     VirusTotal API key. (Use --apikey parameter.)\n"
             << "  --apikey KEY     - sets the API key for VirusTotal\n"
+            << "  --keyfile FILE   - read the API key for VirusTotal from the file FILE.\n"
+            << "                     This way the API key will not appear in the process list\n"
+            << "                     and/or shell history. However, the file name can still be\n"
+            << "                     seen, so proper file permissions should be set.\n"
             << "  --max-age N      - specifies the maximum age for retrieved scan reports to\n"
             << "                     be N days, where N is a positive integer. Files whose\n"
             << "                     reports are older than N days will be updated during the\n"
@@ -69,7 +75,7 @@ void showHelp()
 
 void showVersion()
 {
-  std::cout << "scan-tool-cache, version 0.31, 2016-03-05\n";
+  std::cout << "scan-tool-cache, version 0.32, 2016-03-24\n";
 }
 
 int main(int argc, char ** argv)
@@ -191,6 +197,55 @@ int main(int argc, char ** argv)
             return scantool::rcInvalidParameter;
           }
         } //API key
+        else if (param=="--keyfile")
+        {
+          //only one key required
+          if (!key.empty())
+          {
+            std::cout << "Error: API key was already specified!" << std::endl;
+            return scantool::rcInvalidParameter;
+          }
+          //enough parameters?
+          if ((i+1 < argc) and (argv[i+1] != nullptr))
+          {
+            const std::string keyfile = std::string(argv[i+1]);
+            if (!libthoro::filesystem::file::exists(keyfile))
+            {
+              std::cout << "Error: The specified key file " << keyfile
+                        << " does not exist!" << std::endl;
+              /* Technically it's a file error, but let's return "invalid
+                 parameter" here, because the file name parameter is wrong/
+                 invalid.
+              */
+              return scantool::rcInvalidParameter;
+            } //if file does not exist
+            Configuration conf;
+            if (!conf.loadFromFile(keyfile))
+            {
+              std::cout << "Error: Could not load key from file " << keyfile
+                        << "!" << std::endl;
+              return scantool::rcFileError;
+            }
+            if (conf.apikey().empty())
+            {
+              std::cout << "Error: Key file " << keyfile << " does not contain"
+                        << " an API key!" << std::endl;
+              return scantool::rcFileError;
+            }
+            key = conf.apikey();
+            ++i; //Skip next parameter, because it's used as key file already.
+            #ifdef SCAN_TOOL_DEBUG
+            if (!silent)
+              std::cout << "API key was set to \"" << key << "\"." << std::endl;
+            #endif
+          }
+          else
+          {
+            std::cout << "Error: You have to enter a file name after \""
+                      << param <<"\"." << std::endl;
+            return scantool::rcInvalidParameter;
+          }
+        } //API key from file
         //age limit for update of reports
         else if ((param=="--max-age") or (param=="--age-limit"))
         {