+++ source/scan-tool/main.cpp
@@ -31,6 +31,7 @@
 #include <Windows.h>
 #endif
 #include "summary.hpp"
+#include "../Configuration.hpp"
 #include "../Curly.hpp"
 #include "../virustotal/CacheManagerV2.hpp"
 #include "../virustotal/ScannerV2.hpp"
@@ -52,6 +53,10 @@ void showHelp()
             << "  --version        - displays the version of the program and quits\n"
             << "  -v               - same as --version\n"
             << "  --apikey KEY     - sets the API key for VirusTotal\n"
+            << "  --keyfile FILE   - read the API key for VirusTotal from the file FILE.\n"
+            << "                     This way the API key will not appear in the process list\n"
+            << "                     and/or shell history. However, the file name can still be\n"
+            << "                     seen, so proper file permissions should be set.\n"
             << "  --silent         - produce less text on the standard output\n"
             << "  --maybe N        - sets the limit for false positives to N. N must be an\n"
             << "                     unsigned integer value. Default is 3.\n"
@@ -76,7 +81,7 @@ void showHelp()
 
 void showVersion()
 {
-  std::cout << "scan-tool, version 0.31, 2016-03-05\n";
+  std::cout << "scan-tool, version 0.32, 2016-03-24\n";
 }
 
 /* Four variables that will be used in main() but also in signal handling
@@ -235,6 +240,55 @@ int main(int argc, char ** argv)
             return scantool::rcInvalidParameter;
           }
         } //API key
+        else if (param=="--keyfile")
+        {
+          //only one key required
+          if (!key.empty())
+          {
+            std::cout << "Error: API key was already specified!" << std::endl;
+            return scantool::rcInvalidParameter;
+          }
+          //enough parameters?
+          if ((i+1 < argc) and (argv[i+1] != nullptr))
+          {
+            const std::string keyfile = std::string(argv[i+1]);
+            if (!libthoro::filesystem::file::exists(keyfile))
+            {
+              std::cout << "Error: The specified key file " << keyfile
+                        << " does not exist!" << std::endl;
+              /* Technically it's a file error, but let's return "invalid
+                 parameter" here, because the file name parameter is wrong/
+                 invalid.
+              */
+              return scantool::rcInvalidParameter;
+            } //if file does not exist
+            Configuration conf;
+            if (!conf.loadFromFile(keyfile))
+            {
+              std::cout << "Error: Could not load key from file " << keyfile
+                        << "!" << std::endl;
+              return scantool::rcFileError;
+            }
+            if (conf.apikey().empty())
+            {
+              std::cout << "Error: Key file " << keyfile << " does not contain"
+                        << " an API key!" << std::endl;
+              return scantool::rcFileError;
+            }
+            key = conf.apikey();
+            ++i; //Skip next parameter, because it's used as key file already.
+            #ifdef SCAN_TOOL_DEBUG
+            if (!silent)
+              std::cout << "API key was set to \"" << key << "\"." << std::endl;
+            #endif
+          }
+          else
+          {
+            std::cout << "Error: You have to enter a file name after \""
+                      << param <<"\"." << std::endl;
+            return scantool::rcInvalidParameter;
+          }
+        } //API key from file
         else if ((param=="--silent") or (param=="-s"))
         {
           //Has the silent parameter already been set?