+++ source/honeypottery/ChangeLog.md
@@ -1,5 +1,11 @@
 # ChangeLog for honeypottery
 
+## Version 0.0.5 (2016-03-24)
+  - add option to read the API key from a file
+    This way the API key will not appear in the process list and/or shell
+    history. However, the file name can still be  seen, so proper file
+    permissions should be set to avoid that other users can read the API key.
+
 ## Version 0.0.4 (2015-12-06)
   - fix typo in error message